Group backend sync and workspace routes
